Move Nonce/Index to u32 (#3361)

+++ node/primitives/src/lib.rs
@@ -45,7 +45,7 @@ pub type Balance = u128;
 pub type Moment = u64;
 
 /// Index of a transaction in the chain.
-pub type Index = u64;
+pub type Index = u32;
 
 /// A hash of some data used by the chain.
 pub type Hash = primitives::H256;
